Improve batch archetyp creation and moved examples to solutions folder
