--- getstarted/.gitignore
@@ -1,2 +1,0 @@
-**/.sts4-cache
-**/.vscode
