+++ solutions/getstarted/greeter.js
@@ -1,0 +1,3 @@
+function main() {
+    return {payload: 'Welcome to OpenWhisk on OpenShift'};
+}
